+++ xtuner/v1/ops/comm/__init__.py
@@ -1,5 +1,13 @@
 from .all_to_all import all_to_all_single_autograd, ulysses_all_to_all
+from .nvls_agrs import AllGatherManager, ReduceScatterManager, SymmBufferManager
 from .sequence_parallel import split_for_sequence_parallel
 
 
-__all__ = ["all_to_all_single_autograd", "ulysses_all_to_all", "split_for_sequence_parallel"]
+__all__ = [
+    "all_to_all_single_autograd",
+    "ulysses_all_to_all",
+    "split_for_sequence_parallel",
+    "SymmBufferManager",
+    "AllGatherManager",
+    "ReduceScatterManager",
+]
